[AVRO-2106] Split netty/jetty out of avro-ipc
