+++ lang/java/compiler/pom.xml
@@ -82,6 +82,10 @@
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-surefire-plugin</artifactId>
         <configuration>
+          <!-- some tests hang if not run in a separate JVM -->
+          <forkCount>1</forkCount>
+          <reuseForks>false</reuseForks>
+          <parallel>none</parallel>
           <systemPropertyVariables>
             <test.idl.dir>${project.basedir}/src/test/idl</test.idl.dir>
           </systemPropertyVariables>
@@ -136,16 +140,6 @@
             </configuration>
           </execution>
         </executions>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <configuration>
-          <!-- some tests hang if not run in a separate JVM -->
-          <forkCount>1</forkCount>
-          <reuseForks>false</reuseForks>
-          <parallel>none</parallel>
-        </configuration>
       </plugin>
     </plugins>
     <pluginManagement>
