+++ lang/java/ipc-jetty/pom.xml
@@ -27,31 +27,23 @@
     <relativePath>../</relativePath>
   </parent>
 
-  <artifactId>avro-ipc</artifactId>
+  <artifactId>avro-ipc-jetty</artifactId>
 
-  <name>Apache Avro IPC</name>
+  <name>Apache Avro IPC Jetty</name>
   <url>http://avro.apache.org</url>
-  <description>Avro inter-process communication components</description>
+  <description>Avro inter-process communication components using Jetty</description>
   <packaging>bundle</packaging>
 
   <properties>
     <osgi.import>
-      !org.apache.avro.ipc*,
+      !org.apache.avro.ipc.jetty*,
       org.apache.avro*;version="${project.version}",
-      javax.servlet*;resolution:=optional,
-      org.apache.velocity*;resolution:=optional,
       *
     </osgi.import>
-    <osgi.export>org.apache.avro.ipc*;version="${project.version}"</osgi.export>
-    <osgi.dynamic.import>org.apache.avro.ipc.jetty,org.apache.avro.ipc.netty</osgi.dynamic.import>
+    <osgi.export>org.apache.avro.ipc.jetty*;version="${project.version}"</osgi.export>
   </properties>
 
   <build>
-    <resources>
-      <resource>
-        <directory>src/main/velocity</directory>
-      </resource>
-    </resources>
     <plugins>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
@@ -60,39 +52,9 @@
           <!-- some tests hang if not run in a separate JVM -->
           <forkCount>1</forkCount>
           <reuseForks>false</reuseForks>
-           <parallel>none</parallel>
+          <parallel>none</parallel>
         </configuration>
       </plugin>
-      <plugin>
-        <groupId>${project.groupId}</groupId>
-        <artifactId>avro-maven-plugin</artifactId>
-        <version>${project.version}</version>
-        <executions>
-          <execution>
-            <id>schemas</id>
-            <phase>generate-sources</phase>
-            <goals>
-              <goal>schema</goal>
-              <goal>protocol</goal>
-              <goal>idl-protocol</goal>
-            </goals>
-            <configuration>
-              <excludes>
-                <exclude>**/mapred/tether/**</exclude>
-                <exclude>org/apache/avro/data/Json.avsc</exclude>
-              </excludes>
-              <stringType>String</stringType>
-              <createOptionalGetters>true</createOptionalGetters>
-              <!--<gettersReturnOptional>true</gettersReturnOptional>-->
-              <sourceDirectory>${parent.project.basedir}/../../../../share/schemas/</sourceDirectory>
-              <outputDirectory>${project.build.directory}/generated-sources/java</outputDirectory>
-              <testSourceDirectory>${parent.project.basedir}/../../../../share/test/schemas/</testSourceDirectory>
-              <testOutputDirectory>${project.build.directory}/generated-test-sources/java</testOutputDirectory>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-jar-plugin</artifactId>
@@ -114,11 +76,23 @@
       <version>${project.version}</version>
     </dependency>
     <dependency>
+        <groupId>${project.groupId}</groupId>
+        <artifactId>avro-ipc</artifactId>
+        <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>${project.groupId}</groupId>
       <artifactId>avro</artifactId>
       <version>${project.version}</version>
       <type>test-jar</type>
       <scope>test</scope>
+    </dependency>
+    <dependency>
+        <groupId>${project.groupId}</groupId>
+        <artifactId>avro-ipc</artifactId>
+        <version>${project.version}</version>
+        <type>test-jar</type>
+        <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
@@ -127,29 +101,14 @@
       <scope>test</scope>
     </dependency>
     <dependency>
-        <groupId>com.fasterxml.jackson.core</groupId>
-        <artifactId>jackson-core</artifactId>
+      <groupId>org.eclipse.jetty</groupId>
+      <artifactId>jetty-server</artifactId>
     </dependency>
     <dependency>
-        <groupId>com.fasterxml.jackson.core</groupId>
-        <artifactId>jackson-databind</artifactId>
+        <groupId>org.eclipse.jetty</groupId>
+        <artifactId>jetty-servlet</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.apache.velocity</groupId>
-      <artifactId>velocity-engine-core</artifactId>
-    </dependency>
-    <dependency>
-        <groupId>javax.servlet</groupId>
-        <artifactId>javax.servlet-api</artifactId>
-        <scope>provided</scope>
-    </dependency>
-    <!-- javax.annotation-api dependency must be explicit for forward compatibility (Java > 9) -->
-    <dependency>
-      <groupId>javax.annotation</groupId>
-      <artifactId>javax.annotation-api</artifactId>
-      <version>1.3.2</version>
-    </dependency>
-
   </dependencies>
 
 </project>
+
