--- lang/java/ipc/src/test/java/org/apache/avro/TestProtocolNetty.java
@@ -1,47 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.avro;
-
-import java.net.InetSocketAddress;
-
-import org.apache.avro.ipc.NettyServer;
-import org.apache.avro.ipc.NettyTransceiver;
-import org.apache.avro.ipc.Responder;
-import org.apache.avro.ipc.Server;
-import org.apache.avro.ipc.Transceiver;
-
-/**
- * Protocol test with Netty server and transceiver
- */
-public class TestProtocolNetty extends TestProtocolSpecific {
-  @Override
-  public Server createServer(Responder testResponder) throws Exception {
-    return new NettyServer(responder, new InetSocketAddress(0));
-  }
-
-  @Override
-  public Transceiver createTransceiver() throws Exception{
-    return new NettyTransceiver(new InetSocketAddress(server.getPort()), 2000L);
-  }
-
-  @Override
-  protected int getExpectedHandshakeCount() {
-    return REPEATING;
-  }
-}
