+++ lang/java/pom.xml
@@ -40,6 +40,7 @@
 
     <hadoop2.version>2.7.3</hadoop2.version>
     <jackson.version>2.9.8</jackson.version>
+    <servlet-api.version>3.1.0</servlet-api.version>
     <jetty.version>9.4.14.v20181114</jetty.version>
     <jopt-simple.version>5.0.4</jopt-simple.version>
     <junit.version>4.12</junit.version>
@@ -86,6 +87,8 @@
     <module>compiler</module>
     <module>maven-plugin</module>
     <module>ipc</module>
+    <module>ipc-jetty</module>
+    <module>ipc-netty</module>
     <module>trevni</module>
     <module>tools</module>
     <module>mapred</module>
@@ -457,6 +460,11 @@
           <groupId>org.eclipse.jetty</groupId>
           <artifactId>jetty-servlet</artifactId>
           <version>${jetty.version}</version>
+      </dependency>
+      <dependency>
+          <groupId>javax.servlet</groupId>
+          <artifactId>javax.servlet-api</artifactId>
+          <version>${servlet-api.version}</version>
       </dependency>
       <dependency>
         <groupId>io.netty</groupId>
